improves test helpers to build bundles and pkgs
- adds helper to build provisioner pkgs
- adds support for provisioner pkgs in the bundles
- fixes core rpm build for centos8
- adds logic in fpm to allow conditional skip of yum clean
- makes centos 7.8.2003 a default for testing

+++ images/docker/setup_fpm.sh
@@ -19,7 +19,9 @@
 
 set -eux
 
-yum install -y ruby-devel gcc make rpm-build rubygems python36
+NO_YUM_CLEAN="${NO_YUM_CLEAN:-}"
+
+yum install -y ruby-devel gcc make rpm-build rubygems python36 git
 
 # issues with pip>=10:
 # https://github.com/pypa/pip/issues/5240
@@ -28,4 +30,6 @@ python3 -m pip install -U pip setuptools
 
 gem install --no-ri --no-rdoc rake fpm
 
-rm -rf /var/cache/yum
+if [[ -z "$NO_YUM_CLEAN" ]]; then
+    rm -rf /var/cache/yum
+fi
